requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
openai>=1.0.0
anthropic>=0.30.0
requests>=2.31.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
